split citrine.storage.container.CollectionGroup into new module (citrine.storage.group), remove code for Collection and CollectionGroup from citrine.storage.container, move tests for Collection and CollectionGroup from container unit test file into appropriate unit test files
